Fixing the issue with subsetting LandSat GeoTIFF files
Subsetting of LandSat GeoTIFF files failed because the projection of
LandSat files is different from lat/lon values usually given as input.

Therefore, a function was added to transform the lat/lon values to
the projection of LandSat data (derived automatically)

Furthermore, subsetting function was adjusted to accept the transformed
lat/lon values as local variables.

Reported-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>
Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>