Correcting extent extraction from LandSat GeoTIFFs
Since converting the extents of input Shapefiles to the Landsat's
projection value resulted in spatially accurate GeoTIFFs, another
solution was implemented to: 1) reproject the input ESRI Shapefile into
the projection of the Landsat rasters and 2) extract the extents from
the reprojected ESRI Shapefile. This results in accurate spatial subset
of the Landsat landcover values.

Furthermore, the `extract_extent_shapefile()` function has changed in a
way that could be provided with the source projection where the extents
are desired. If not provided, the source projection of the ESRI
Shapefile input data is considered to extract the extents.

Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>
Reported-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>