Iss14 (#30)
* Fixing the issue with subsetting LandSat GeoTIFF files

Subsetting of LandSat GeoTIFF files failed because the projection of
LandSat files is different from lat/lon values usually given as input.

Therefore, a function was added to transform the lat/lon values to
the projection of LandSat data (derived automatically)

Furthermore, subsetting function was adjusted to accept the transformed
lat/lon values as local variables.

Reported-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>
Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>

* bumping version to 0.1.4

* removing the duplicate line for specifying longitude limits variable

* Correcting extent extraction from LandSat GeoTIFFs

Since converting the extents of input Shapefiles to the Landsat's
projection value resulted in spatially accurate GeoTIFFs, another
solution was implemented to: 1) reproject the input ESRI Shapefile into
the projection of the Landsat rasters and 2) extract the extents from
the reprojected ESRI Shapefile. This results in accurate spatial subset
of the Landsat landcover values.

Furthermore, the `extract_extent_shapefile()` function has changed in a
way that could be provided with the source projection where the extents
are desired. If not provided, the source projection of the ESRI
Shapefile input data is considered to extract the extents.

Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>
Reported-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>

---------

Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>